Fix typo in Meals db and update number_of_servings to float in MealFoods db

+++ db/migrate/20200805015444_create_meal_foods.rb
@@ -3,7 +3,7 @@ class CreateMealFoods < ActiveRecord::Migration[6.0]
     create_table :meal_foods do |t|
       t.references :meal, null: false, foreign_key: true
       t.references :food, null: false, foreign_key: true
-      t.integer :number_of_servings
+      t.float :number_of_servings
       t.integer :calories
 
       t.timestamps
